flat-white: Fix non-Rmd rchunk code styles without breaking inline code
It's a bit hacky, but both inline `code` and normal code chunks are given .ace_support.ace_function class. I added a selector that tries to catch code chunks, which are "inline code" except that they definitely are either the first one or the first child in the line is an invisible character (.ace_invisible).

The unforunate side-effect is that the first backtick of an inline code chunk won't be highlighted correctly.